backend/requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==4.2.29
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
idna==3.11
PyJWT==2.12.1
python-dotenv==1.2.1
requests==2.32.4
sqlparse==0.5.5
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
